fix: correct type definitions
`equals` should take a value of the same type but according to the
specification, if the value is not of the same type, returning false is
recommended. `equals` should therefore not throw if the value is not of
the same type.

`ord` is the same case a `equals` but here the definition was actually
incorrect.